move shared methods into utils
