+++ validation/testutils/mock_ethclient.go
@@ -8,10 +8,8 @@ import (
 	"github.com/ethereum/go-ethereum/crypto"
 )
 
-// MockEthClient is a mock that implements EthCaller for testing.
 type MockEthClient struct {
-	// Setup response maps keyed by method 4-byte signature + address, or
-	// just by method signature if that's simpler for your tests.
+	// Response map keyed by method 4-byte signature
 	Responses map[string][]byte
 	Err       error
 }
@@ -23,7 +21,7 @@ func (m *MockEthClient) CallContract(ctx context.Context, msg ethereum.CallMsg, 
 	return m.Responses[string(msg.Data)], nil
 }
 
-// Helper to get the 4-byte method ID (like what getBytes does internally)
+// Helper to get the 4-byte method ID
 func MethodID(method string) []byte {
 	return crypto.Keccak256([]byte(method))[:4]
 }
