+++ validation/utils.go
@@ -11,12 +11,66 @@ import (
 
 	"github.com/ethereum-optimism/optimism/op-service/retry"
 	"github.com/ethereum-optimism/superchain-registry/superchain"
+	"github.com/ethereum-optimism/superchain-registry/validation/standard"
 	"github.com/ethereum/go-ethereum"
+	"github.com/ethereum/go-ethereum/common"
+	"github.com/ethereum/go-ethereum/crypto"
+	"github.com/ethereum/go-ethereum/ethclient"
 	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
 )
 
 type EthClient interface {
 	CallContract(ctx context.Context, msg ethereum.CallMsg, blockNumber *big.Int) ([]byte, error)
+}
+
+func getAddressFromConfig(chainID uint64, contractName string) (superchain.Address, error) {
+	if common.IsHexAddress(contractName) {
+		return superchain.Address(common.HexToAddress(contractName)), nil
+	}
+
+	contractAddress, err := superchain.Addresses[chainID].AddressFor(contractName)
+
+	return contractAddress, err
+}
+
+func getAddressFromChain(method string, contractAddress superchain.Address, client EthClient) (superchain.Address, error) {
+	addr := (common.Address(contractAddress))
+	callMsg := ethereum.CallMsg{
+		To:   &addr,
+		Data: crypto.Keccak256([]byte(method))[:4],
+	}
+
+	// Make the call
+	callContract := func(msg ethereum.CallMsg) ([]byte, error) {
+		return client.CallContract(context.Background(), msg, nil)
+	}
+	result, err := Retry(callContract)(callMsg)
+	if err != nil {
+		return superchain.Address{}, err
+	}
+
+	return superchain.Address(common.BytesToAddress(result)), nil
+}
+
+var checkResolutions = func(t *testing.T, r standard.Resolutions, chainID uint64, client *ethclient.Client) {
+	for contract, methodToOutput := range r {
+		contractAddress, err := getAddressFromConfig(chainID, contract)
+		require.NoError(t, err)
+
+		for method, output := range methodToOutput {
+			want, err := getAddressFromConfig(chainID, output)
+			require.NoError(t, err)
+
+			got, err := getAddressFromChain(method, contractAddress, client)
+			require.NoErrorf(t, err, "problem calling %s.%s (%s)", contract, method, contractAddress)
+
+			// Use t.Errorf here for a concise output of failures, since failure info is sent to a slack channel
+			if want != got {
+				t.Errorf("%s.%s = %s, expected %s (%s)", contract, method, got, want, output)
+			}
+		}
+	}
 }
 
 // isBigIntWithinBounds returns true if actual is within bounds, where the bounds are [lower bound, upper bound] and are inclusive.
